Fix database type & minor change display in orderhistory

+++ database/migrations/2014_10_12_000000_create_users_table.php
@@ -21,7 +21,7 @@ class CreateUsersTable extends Migration
             $table->timestamp('email_verified_at')->nullable();
             $table->string('password');
             $table->string('fullName');
-            $table->integer('gender');
+            $table->tinyInteger('gender');
             $table->string('phone')->unique();
             $table->string('address');
             $table->rememberToken();
